Move and rename minimizer specific modules
