+++ zfit/core/__init__.py
@@ -1,2 +1,2 @@
-from . import (baseobject, basepdf, basemodel, basefunc, data, interfaces, integration, minimizer, math, loss, tfext,
+from . import (baseobject, basepdf, basemodel, basefunc, data, interfaces, integration, math, loss, tfext,
                sample, limits, operations, parameter, )
