+++ zfit/minimizers/fitresult.py
